Remove the deep extend method that was written for the graphtool.  The jsxgraph library has a merge method that does this, so just use that.
If a graphtool graph container object does not have width or height when
the window is resized, then delay resizing the graph until the container
does have width and height.  This can happen, for example, if the
graphtool is inside a closed scaffold when the window is resized.

Change from using the title attribute for the tooltip on the divs
containing the tool buttons, to using the data-bs-title attribute.  If
the title attribute is used then bootstrap adds an aria-label attribute
to the div.  That attribute is not valid on these divs.  This means that
the buttons themselves can not be aria-labelledby the div since they
don't have a visible text (as far as the browser is concerned).  So the
buttons now directly have the aria-label whose value is the same as that
of the data-bs-title for the div.

Fix the nestedClose value of the GraphTool context.  That should
correspond to the nestedOpen value.  Although, it didn't seem to cause a
problem.