Fix the adjustment of unit answers for comparison.
Instead of adjusting the correct answer by the correct factor over the
student factor, adjust the student answer the student factor over the
correct factor.  Furthermore, only do that temporarily while the
comparison is made, and then revert it.

The adjustment is made at a much lower level now.  Instead of doing it
in the `cmp_parse` method, do it in the `compare` method which is the
overload method for the `<=>` operator.  That is the lowest level of
evaluation in the MathObjects answer processing sequence.  This means
that it is after the display values for the student answers are set.

Some additional unit tests for this have been added in
t/units/basic_parser.t.  Also the "todo" tests that are related to this
are now actual tests.

This fixes issue #740.