Add the parserRadioMultiAnswer.pl macro.
This macro in essence combindes the parserRadioButtons.pl and
parserMultiAnswer.pl macros.  It basically allows the creation of
answers that are radio button answers where each radio button has child
answers.

The point of this type of answer is for problems where different types
of answers are expected in different cases.  For example, a problem that
asks a student to solve a system of linear equations.  The possible
answers are of course a single solution, an infinite number of
solutions, or no solutions.  So using this macro you could create an
answer that has three radio buttons.  If the solution is a single
solution, then the single solution radio part would have answer boxes in
it for the student to enter the solution.  If there are an infinite
number of solutions, then the radio button for that part would have
answer boxes to enter the parameterization for the solution set.  If
there are no solutions, then that part would not have answer boxes (no
need).  So the student would see something like:

* A. The unique solution is x = ___, y = ___.
* B. There are an infinite number of solutions parameterized by x = ___, y = ___.
* C. There are no solutions.