Add a macro for multiple choice checkbox answers.
This is intended to be a replacement for using PGchoicemacros.pl with
new_checkbox_multiple_choice.

In order for this to work when used with a MultiAnswer object with
singleResult true, the NAMED_ANS_CHECKBOX and NAMED_ANS_CHECKBOX_OPTION
methods need to be updated to match the NAMED_ANS_RADIO and
NAMED_ANS_RADIO_EXTENSION methods.

In addition, this fixes repeated ids on checkboxes created via
NAMED_ANS_CHECKBOX_OPTION, and makes the coloring of checkboxes with
incorrect and incorrect colors work as it does for almost all other
answers.