Fix and deprecate unionInclude.pl
Make the unionInclude.pl macro work again.  `fileName` is no longer in
the environment.  Instead of having `fileName` and `probFileName` in the
environment that were both always the same thing, there is now only
`probFileName`.

However, this macro should not be used anymore, so it is also moved to
deprecated.