Rework the mqeditor toolbar postioning and behavior
The toolbar is absolutely positioned relative to the position of the
answerQuill input on the page that it corresponds to.  This means that
the toolbar will stay with the input when the page scrolls.  If the
problem is rendered in an iframe, then the toolbar is still always to
the right of the input.  Care is taken to ensure that the toolbar
doesn't go off the top or bottom of the page (unless the page is too
short to accommodate the toolbar height -- see below).

The mqeditor toolbar is now scrollable when in short pages.  The toolbar
does not collapse into multiple columns, and is always a single column.
So it doesn't intrude into the problem when the page is short anymore
than its single column 53px width (plus a 10 pixel gap to the right of
the toolbar).

In order for the toolbar to be scrollable, it needs to be able to attain
focus.  Otherwise it will disappear when the user tries to scroll it
using the scrollbar.  So the toolbar is inserted into the page
immediately after the MathQuill input.  This means that if the MathQuill
input has focus and the tab button is pressed, then focus will go to the
first button of the toolbar (and then on through the buttons when tab is
pressed additional times).  If you tab all the way through the buttons,
then focus will continue in the page to the next place it would have
gone if the toolbar wasn't there.  Note that the toolbar is not in the
reverse tab order when using shift-tab from somewhere after the
MathQuill input (since it doesn't exist at that time).  This new tab
order approach is probably more accessible in any case, since the
toolbar now can be used with the keyboard.

Finally, if you don't want the toolbar to open for you at all, then type
alt-t when any MathQuill input has focus.  The toolbar will hide, and
will not open again for any MathQuill input.  If you type alt-t again in
any MathQuill input, then the toolbar will open and thereafter will
continue to open for all MathQuill inputs.  This setting is saved in
local storage, and so will persist for any page on the site. I.e., alt-t
typed in any MathQuill input toggles whether the toolbar is enabled or
not.