Update the quickMatrixEntry.pl macro
This removes the dependence on jQuery and jquery-ui.

This also adds an ans_rule method for PGML usage.  The POD documentation
that was added explains only how to use the macro in this way.

There is deprecated code for backwards compatibility, although I would
gladly remove that.  The only known problem that uses this macro is in
the pg repository.  It is the file
`t/matrix_tableau_tests/tableau_javascript.pg`.  I would also like to
delete that file.