Add the capability to set the radio input value to parserRadioButtons.pl
The values for the choices are the student "original" answers which are
displayed in the past answers table (or answer log).  Currently for this
macro those are always B0, B1, etc.  This changes that so those can be
customized to show more informative "original" answers in the results
table.  Note that if using a custom checker, these are also the values
that you would need to use for comparison there.

These values can be specified in two ways.  One extends to the previous
method of providing a label with a choice. For example, a choice could
have been give as { label => text }.  To give a value with that now use
{ label => [ text, value ] }.  For a full example:

```
$radio = RadioButtons(
	[
		{ A => [ 'answer 1', 'always first answer' ] },
		[
			{ B => [ 'answer 2', 'random order second answer' ] },
			{ C => [ 'answer 3', 'random order third answer' ] }
		]
		{ D => [ 'answer 4', 'always last answer' ] }
	],
	1,
	displayLabels => 1
);
```
Note that with this approach the labels, text, and value are all
randomized together and everything is shown in that random order.
So in the problem you will get for instance
* A. answer 1 (the value of this radio will be 'always first answer')
* C. answer 3 (the value of this radio will be 'random order third answer')
* B. answer 2 (the value of this radio will be 'random order second answer')
* D. answer 4 (the value of this radio will be 'always last answer')
Of course B and C can be switched.

The other way of specifying the values is to give a reference to an
array containing values for each of the choices.  For example,
```
$radio = RadioButtons(
	[
		'answer 1',
		[
			'answer 2',
			'answer 3',
		]
		'answer 4',
	],
	1,
	labels => 'ABC',
	values => [
		'always first answer',
		'random order second answer',
		'random order third answer',
		'always last answer'
	]
);
```
Note that with this approach the values are still randomized with the
choices, but the labels are not.  So in the problem you will get for
instance
* A. answer 1 (the value of this radio will be 'always first answer')
* B. answer 3 (the value of this radio will be 'random order third answer')
* C. answer 2 (the value of this radio will be 'random order second answer')
* D. answer 4 (the value of this radio will be 'always last answer')
Of course the text for B and C can be switched and the values with them,
but the labels B and C will always be in the order shown.

Of course one should choose the values to not give the correct answer
away.

This also moves the "uncheckable" javascript out of the macro and into a
separate javascript file loaded via ADD_JS_FILE.  The javascript is
updated to not pollute the global namespace.

Note that the NAMED_ANS_RADIO_EXTENSION of PGbasicmacros.pl needed to be
extended to allow setting of the input `id` by the caller.  Previously
"${name}_$value" was used.  The values set by the author could make
those invalid id's.

Both that method and the NAMED_ANS_RADIO method also needed to be
extended to allow the caller to add attributes to the input.  This is to
that data attributes could be added that are used by the javascript.
Previosly this macro used a regular expression substitution to add an
`onclick` handler.  Really what is needed is for the PGbasicmacros.pl
macro to have a generic method to generate html tags with attributes.