Add a tag helper method to PGbasicmacros.pl
This method is similar in usage to the Mojolicious::Plugins::TagHelpers
`tag` method.

Currently it is used in NAMED_ANS_RADIO, NAMED_ANS_RADIO_EXTENSION,
NAMED_ANS_CHECKBOX, and NAMED_ANS_CHECKBOX_OPTION, but should be used in
the other answer rule methods in that file, and perhaps many other
places.