Implement the `values` option for parserRadioMultiAnswer.pl.
For this macro the only way to pass the values is to use the `values`
option.  The hash method is not a possibility for this macro.  In
addition, this macro does not randomize the order of radio buttons.  So
setting the values may not be as important.

Also, correct some POD typos in parserRadioMultiAnswer.pl and use the
`tag` method for html generation.