Make it possible to pass options to the cmp method with PGML
Currently passing options to the cmp method using PGML answer methods is
not possible.  So instead you have to resort to using the basic PG
approach of calling ans_rule or ans_array, and then calling the cmp
method yourself.

This makes it so that you can pass those options using the PGML answer
syntax.  This can be done for example with:
```
$ans = Matrix([ [ 1, 2, 3 ], [ 4, 5, 6 ]]);
$ans_cmp_opts = {
	checker => sub { ... }
};

BEGIN_PGML
Matrix answer: [_]*{answer=>$ans, cmp_options=>$ans_cmp_opts}
END_PGML
```
You can also use:
`Matrix answer: [_]*{$ans}{5}{NEW_ANS_NAME()}{$ans_cmp_opts}`
Unfortunately, the order is answer, width, name, cmp_options, and that
can't be changed for compatibility reasons.  So you  have to pass an
answer name with this approach which should always be `NEW_ANS_NAME()`
or a scalar variable whose value is the result of calling that method.

This new option is most useful for using a custom checker with a
multitude of answer types like matrices (as demonstrated above).
parserRadioButtons.pl is another case where this could be used, and I am
sure there are many others.