Remove CRLF line endings when a PG file is loaded.
Currently a pg file that has CRLF line endings fails to render.  This
fixes that by removing such line endings in the default_preprocess_code
method.

An alternate approach that would also work would be to do this when the
translator opens a file in the source_file method.  However, this would
not catch cases when a source string that contains CRLF line endings is
provided instead of a source file.  That may not be a real concern
though?

I am not exactly sure what changed that caused files with CRLF line
endings to fail.  They work for PG 2.17, but do not for the develop
branch.