Fix the order of checkbox answers.
For this to work extended response group answers need to preserve order.
So instead of a hash these extended response group answers are now
arrays of label/value pairs.  Note that the label is really the input
value, and the value is either 'CHECKED' or the empty string.  Previosly
the hash key was the input value, and its value was 'CHECKED' or the
empty string.  Note that when the translators processes these answers it
also no longer converts the answer into a null byte separated string.
It is instead converted into a flat array.  PG is already set up to
handle this.  This is just one small part of eliminating the null byte
hack.

Also some clean up of annoying old code in the translator.  The safety
filter has been removed.  It has been stubbed with the "null filter"
since 2003.  The "null filter" just returns the input.  No need for
that.  The inputs ref argument to the translator process_answers method
has been removed.  That argument has not used since at least 2005.