Add unicode "degree Celsius", "degree Fahrenheit", and "Kelvin" units.
Add the unit K without the degree symbol, and make it the fundamental
unit for Kelvin.  It is the correct SI unit.  Not the one with the
degree symbol.

Also switch to using the unicode escape sequence rather than the pasted
unicode character.  The pasted unicode characters can be problematic in
text editors.