Add a basic alias system for units.
Aliases can be added for a unit via the 'aliases' key.  The value should
be an array of aliases.  These are added as keys to the %known_units
hash and the 'aliases' key deleted at compile time.

Also fix some perlcritic issues in the file, and clean it up a bit.

The "degree Kelvin" units have been removed from the help files, but not
from the actual units allowed (for now).

Also change Centrigrade to Celsius in the help files.  What is a
Centrigrade?  Also, a Centigrade is not correct either.  It has been
Celsius since 1948.