Rewrite unionTables.pl to replace deprecated html attributes.
The deprecated attributes are replaced with css. Some is inline, and
some is added via css classes and styles defined in the
htdocs/js/UnionTables/union-tables.scss file.  One thing that works
correctly again that didn't before is borders.  The reason that borders
were not working is because of bootstrap styling which overrides the
deprecated border attribute.

In addition the POD is rewritten in a nicer format.

Although this fixes the deprecated attributes and the borders now appear
correctly, the only real conclusion that can be made in looking at the
way this macro works is that this macro needs to be deprecated.  The
likelyhood of invalid html being generated by this macro is quite high.
Due to its design, there is no way to validate the usage by problem
authors.  If the macro's TableSpace or TableLine methods are used, the
html that is generated is almost guaranteed to be invalid unless the
table has precisely 10 columns.

In addition, the usage of this macro in OPL problems is frequently for
layout.  Using html tables for layout is always wrong.  To be fair, this
used to be pretty much the only way to do these types of layouts.  Now
it is just wrong, and there are better ways to do such layouts.