Add a numberLine graphing mode to the graphtool.
An interval tool is added that is designed to work with this graphing
mode.  Note that the extreme left and right ends of the graph are
considered points at infinity.  So both bounded and unbounded intervals
can be graphed with this tool.

Endpoints can be made to be include or excluded with an include/exclude
tool that works like the previous solid/dash tool for curves.  It is a
bit different though as it applies per point, not for the entire object.

Some examples of the answer format for this graph object are
"{interval,(-4,5]}", "{interval,(-infinity,10)}", or
"{interval,[5,infinity)}".

If the option useBracketEnds is set to 1, then the endpoints are
displayed as parentheses and brackets instead of open and closed dots
(which is the default).