Fix knowls in lists.
When knowls are added in a list, instead of adding an `li` tag and
inserting that into the list, add a `div` tag and insert that inside the
existing `li` tag.  This results in the same appearance, is valid html,
and does not affect numbering inside an ordered list.

This fixes issue #794.