Add a coordinateHintsType option to the graph tool.
This option changes the way coordinate hints are shown.  By default the
coordinates are displayed as decimal numbers accurate to five decimal
places.  If this is set to 'fraction', then those decimals will be
converted and displayed as fractions.  If this is set to 'mixed', then
those decimals will be converted and displayed as mixed numbers.  For
example, if the snapSizeX is set to 1/3, then what would be displayed as
4.66667 with the default 'decimal' setting, would be instead be
displayed as 14/3 with the 'fraction' setting, and '4 2/3' with the
'mixed' setting.  Note that these fractions are typeset by MathJax.