Improved non-recursive version of contFrac in contextFraction.pl.
This version uses a loop instead of a recursive function.  Benchmarking
shows it is at least twice as fast.  It also should use less memory
overhead as it does not store the steps in constructing the continued
fraction in arrays.  It only keeps what is needed to continue.  That is
the last two values of the convergent numerator and denominator.

A basic unit test is added to check that this does the right thing.