Add a dynamic help system.
Help messages are displayed as the user interacts with the graph tool.

The bootstrap tooltips have been removed.  Instead those use the new
help message box.  The tooltip messages are also bit more descriptive of
the tool.

Instead of bootstrap modal dialogs for confirming object deletion or
clearing the board, a sort of dialog in the content of the message box
is used.

Button styles were added to replace the bootstrap button styles.

This makes the graph tool no longer dependent on bootstrap.