Update the drag handlers for the graphtool graph objects.
This takes advantage of the keyboard motion events that are now sent by
JSXGraph to drag and move handlers as of version 1.5.0.  Now the
graphtool drag and move handlers that previously only handled pointer
events also handle keyboard motion events.

This also fixes an issue on touch screen devices where it is possible to
drag objects off the board.