Update mathview to not use jQuery and to use modern javascript.
This also fixes a bug in the current code.  The math in the math view
popover was no longer being typeset.  I suspect this was caused by the
recent bootstrap update and a change in the way that popovers work.  The
math view code was attempting to typeset the code when the button is
clicked.  That triggers the opening of the popover, but since it is not
yet open, the typesetting does nothing.  This moves the typesetting
request into the show.bs.popover event handler that bootstrap triggers
when the popover is added to the DOM.

The typeahead code that I had previously commented out is now removed.
I looked into finding a library that replaces bootstrap twitter's
typeahead functionality, but was unable to find one that has the
features that were utilized by math view.  In fact the latest version of
bootstrap twitter's typeahead (finalized 8 years ago) had already
removed some of the functionality that math view depended on.  If
someone would like to investigate further, feel free.