Give essay answers a MathQuill equation editor.
This adds a button to the essay textarea that when clicked opens a
collapse below the text area.  The collapse contains a MathQuill entry
box and an insert button.  The user can type equations in the MathQuill
entry, and then click "Insert" to inject latex wrapped in `\(` and `\)`
into the essay textarea at the current cursor location.

This is perhaps not quite as nice as the usual MathQuill answer box for
regular input type text answers, but at least provides something for
essay type answers.  This is vaguely similar to what MathView does for
essay type answers.

Note that PG already adds hidden latex MaThQuIlL_AnSwEr00?? blanks for
essay answers.  Previously those were ignored as the corresponding
AnSwEr00?? blank is not of type text and doesn't have the codeshard
class.  That hidden field is used to identify the textareas in
combination with the latexentryfield class that essay answers have.
However, unlike regular text type inputs nothing is actually saved in
the hidden latex field.