Add a TO_JSON method to Value objects.
This method is essentially an alias of the `string` method that Value
objects already have.  TO_JSON is called by JSON->encode when
convert_blessed is true.  So if you call
JSON->convert_blessed->encode($value) where $value is a Value object or
where $value contains an array entry or hash key that is a Value object,
these Value objects will be properly converted into a string in the
output.