Replace bin/run-perltidy with bin/run-perltidy.pl
The primary advantage of this perl script over the previous shell script
is that it used File::Find instead of relying on the extglob, globstar,
and nullglob shell options.  Those options are not supported by all
shells.

If this script is run with no arguments, then it does the same thing
that the previous shell script did, and runs perltidy on all .pm, .pl,
and .t files in the PG_ROOT directory.

If file names are passed on the command line, then perltidy is only
executed on the passed filenames.  Another advantage of this script over
the previous shell script.

The script enforces the use of the .perltidyrc file in the PG_ROOT
directory.  It also makes the default for -bext to be '/', which means
that backup files are created with the .bak extension, and deleted if no
errors occur.  That can be overridden if backups are desired.  Other
that those things, all options accepted by perltidy are also accepted by
this script.

One other thing that this script does is enforces that the version of
perltidy used is version 20220613 or newer.  Older versions have an
issue with the -vxl='q' option.