Rewrite third party asset handling.
This removes the third-party-assets.json file, and the contents of that
file are no longer added to the static-assets.json file.  Instead the
frontend (webwork2 or otherwise) should load the package.json file and
use that to determine how to serve any file located in node_modules.

Note that a version tag is added to mathquill in the package.json file
to lock the specific commit of mathquill to use.  The tag for PG-2.18 is
WeBWorK-2.18.  Previously a tag for 2.17 was added.  If a hotfix is
needed for mathquill, then we will append a subversion to ensure that
browser caching is avoided.