Add an essay field preview button and restructure the equation editors a bit.
The essay preview button is added via the htdocs/js/apps/Essay/essay.js
javascript.  This javascript is now always added to the problem page
(via load_js in PG.pl).  This preview button is added below essay
answers regardless of the entry assist method.

The restructuring is to accomodate the usage of these editors for the
webwork2 manual problem grader comment boxes, but I think is better in
general.  The button attached to the top right corner of the textarea
doesn't really look that great.

For the MathQuill equation editor, the button that activates the
equation editor is now placed below the essay textarea and to the left
of the new preview button.  In addition, the "Insert" and "Clear"
buttons of the equation editor are placed in the card footer below the
MathQuill input instead of to the right of it.

For the MathView equation editor, the button that activates the equation
editor is also placed below the essay textarea, but to the right of the
preview button.  Furthermore, those buttons are wrapped in a div that
aligns its content to the right.  This makes it so that when the
equaiton editor popover opens it does not obscure the textarea.  The
fallback placements have been modified so that the popover prefers to go
to the right if there is room, but if not it tries to first go down, and
then up.  It is not allowed to go to the left.  In some extreme
situations (short windows with the button already on the far right) this
will cause the page to be extended to the right to make room for the
popover.

Also for the MathView equation editor, a "Clear" button is added and the
"Insert" and "Clear" buttons are placed right justified below the
equation editor input instead of to the right of it.