Tweak the appearance of the essay equation editor.
Open the equation editor below the equation editor and preview buttons
so the buttons don't move when the equation editor opens.

Remove the top border on the equation editor card footer, and set its
background color to white to match that of the card body.

Make the equation editor card header have the secondary color for its
background color, and the close button in the header white.

Fix the flickering of the toolbar when the "Clear" button is clicked.