Improvements to the sample problems.
All problems are perltidied with a line length limit of 80.

The ManyMultipleChoice.pg problem was rewritten to use parserPopUp.pl
instead of PGchoicemacros.pl.

An alternate matching problem was added that demonstrates an approach to
creating such a problem without PGchoicemacros.pl.  It takes a little
more work to write the problem this way though.  So the original
Matching.pg problem was left.

Some other example problems were added as well.

The `problem-template.mt` template now html escapes code.  Otherwise
html in problems does not appear.  In fact, `<` and `>` in problems were
already causing issues for many of the problems.

The TikZ in sample-problems/DiffCalcMV/ContourPlot.pg and
sample-problems/Precalc/DynamicGraph.pg was improved.  Adding a
background color to the image helps to offset the image from the problem
background.  Also the `$refreshCachedImages` variable was removed.  That
should not be used pretty much at all.  It forces the server to
regenerate the TikZ image everytime.  It does not make the browser
reload images as the comments claimed.  Note that if the problem source
changes a new image will be generated by the server regardless of if
that variable is set.  When students are working the problem you
certainly do NOT want that variable set, as it will be a drain on the
server and is not needed.

Note that all "Precalc" problems have been moved to "Algebra" and the
keywords and such corrected.  Precalculus is not an actual mathematical
subject area.  It is a course.  The contents of a typical precalculus
course are a mix of topics from other areas, usually algebra and
trigonometry.  These problems were all algebra problems.

Also tweak the sidebar.  Use a list-group instead of a nav with
nav-pills.  It just looks a little better.

The sidebar was also broken on narrow screens and that was fixed.