Switch to using Pandoc instead of the Text::MultiMarkdown package.
The Text::MultiMarkdown package is not maintianed, and the last release
was in 2014.  As such it does not support much of modern markdown
syntax.  For example, it does not support triple backtick code blocks.

Pandoc is actively maintained and does support modern markdown (and much
more).  It is actually a separate application, and the Pandoc perl
module is really just a wrapper that makes it possible to use in perl
scripts.