In WeBWorK::PG::Environment get `$pg_root` from the PG_ROOT environment variable.
This is just for correctness as the current code does work.  However,
the only reason it works is because of the fact that `pg_root` is no
longer set in the webwork2 CourseEnvironment.  So what happens is that
the eval statement on lines 64-68 returns undefined and so the `or do`
statement is executed.  So in the end the `$pg_root` is correctly set,
but from the PG_ROOT environment variable set by the Mojolicious
webwork2 app when it starts.  Note in this case it does not use the
system environment variable even if that is set since the app overrides
it with the value of pg_dir in webwork2.mojolicious.yml.