Add a default checker to the parserGraphTool.pl macro.
The default checker is designed to grade the graph based on appearance.
This means that if a student graphs duplicate objects, then the
duplicates are ignored.  Furthermore, if two objects are graphed whose
only difference is that one is solid and the other is dashed (in this
case the dashed object is covered by the solid object and only the solid
object is really visible), then the dashed object is ignored.

Note that fills are only checked by this checker if the other graph
objects are correct.  This is because, without the other elements of the
correct answer, there is not a valid setting to evaluate if the fill is
correct or not.

The variable `$graphToolObjectCmps` is made available for use in custom
checkers, and contains a hash whose keys are the types of the objects
available (line, circle, parabola, fill, point, cubic, quadratic, and
inverval), and whose values are methods that can be called passing one
of these objects (typically a correct answer in a checker).  When one of
these methods is called it will return two methods.  The first method
when called passing a MathObject point will return 0 if the point
satisfies the equation of the object, -1 if the equation evaluated at
the point is negative, and 1 if the equation evaluated at the point is
positive.  The second method when called passing another graph tool
object (for example a student answer) will return 1 if the two objects
are exactly the same, and 0 otherwise.  A second parameter may be passed
and if that parameter is 1, then the method will return 1 if the two
objects are the same ignoring if the two objects are solid or dashed,
and 0 otherwise.  See the added POD for an example of using this in a
custom checker.