Implement workaround for JSXGraph issue with points created off the board.
If a point is created off the board with `snapSizeGrid` true, then the
coordinate that is outside of the board's bounding box range is moved
one snap size unit toward the bounding box range.  See
https://github.com/jsxgraph/jsxgraph/issues/545 for issues on this bug
in JSXGraph.

To fix this points need to be created with `snapToGrid` false (the
default value), and then that attribute set to true afterward.  In order
for this to work the coordinates of the point need to be manually
rounded to the snap size.

This is only done for initial creation of objects (answers or static
objects).  It does not need to be done for objects that are created
dynamically (by student interaction) since those points are forced to be
on the board anyway.  This is also not done for 'point' graph tool
objects, since the correct answer should not be a point off the board
anyway.

This workaround fixes issue #849.